backend - iNEWS/requirements.txt
absl-py==1.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-vite==2.0.2
djangorestframework==3.14.0
docx2txt==0.8
executing==0.9.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
flatbuffers==1.12
fpdf==1.7.2
future==0.18.2
gast==0.4.0
google==3.0.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
jsons==1.6.3
keras==2.9.0
Keras-Preprocessing==1.1.2
langdetect==1.0.9
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
networkx==2.8.5
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF2==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pyvis==0.2.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
snscrape==0.5.0.20230113
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.3.0
stopwords==1.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.3.1
toml==0.10.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
typish==1.9.3
tzdata==2022.7
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
